Changed error codes to negative integers

+++ src/modules/SX127x.cpp
@@ -4,7 +4,7 @@ SX127x::SX127x(Module* mod) {
   _mod = mod;
 }
 
-uint8_t SX127x::begin(uint8_t chipVersion, uint8_t syncWord) {
+int16_t SX127x::begin(uint8_t chipVersion, uint8_t syncWord) {
   // set module properties
   _mod->init(USE_SPI, INT_BOTH);
   
@@ -39,7 +39,7 @@ uint8_t SX127x::begin(uint8_t chipVersion, uint8_t syncWord) {
   }
   
   // set LoRa sync word
-  uint8_t state = SX127x::setSyncWord(syncWord);
+  int16_t state = SX127x::setSyncWord(syncWord);
   if(state != ERR_NONE) {
     return(state);
   }
@@ -47,7 +47,7 @@ uint8_t SX127x::begin(uint8_t chipVersion, uint8_t syncWord) {
   return(ERR_NONE);
 }
 
-uint8_t SX127x::transmit(uint8_t* data, size_t len) {
+int16_t SX127x::transmit(uint8_t* data, size_t len) {
   // check packet length
   if(len >= 256) {
     return(ERR_PACKET_TOO_LONG);
@@ -103,15 +103,15 @@ uint8_t SX127x::transmit(uint8_t* data, size_t len) {
   return(ERR_NONE);
 }
 
-uint8_t SX127x::transmit(const char* str) {
+int16_t SX127x::transmit(const char* str) {
   return(SX127x::transmit((uint8_t*)str, strlen(str)));
 }
 
-uint8_t SX127x::transmit(String& str) {
+int16_t SX127x::transmit(String& str) {
   return(SX127x::transmit(str.c_str()));
 }
 
-uint8_t SX127x::receive(uint8_t* data, size_t len) {
+int16_t SX127x::receive(uint8_t* data, size_t len) {
   // set mode to standby
   setMode(SX127X_STANDBY);
   
@@ -175,7 +175,7 @@ uint8_t SX127x::receive(uint8_t* data, size_t len) {
   return(ERR_NONE);
 }
 
-uint8_t SX127x::receive(String& str, size_t len) {
+int16_t SX127x::receive(String& str, size_t len) {
   // create temporary array to store received data
   char* data = new char[len];
   uint8_t state = SX127x::receive((uint8_t*)data, len);
@@ -189,7 +189,7 @@ uint8_t SX127x::receive(String& str, size_t len) {
   return(state);
 }
 
-uint8_t SX127x::scanChannel() {
+int16_t SX127x::scanChannel() {
   // set mode to standby
   setMode(SX127X_STANDBY);
   
@@ -214,17 +214,17 @@ uint8_t SX127x::scanChannel() {
   return(CHANNEL_FREE);
 }
 
-uint8_t SX127x::sleep() {
+int16_t SX127x::sleep() {
   // set mode to sleep
   return(setMode(SX127X_SLEEP));
 }
 
-uint8_t SX127x::standby() {
+int16_t SX127x::standby() {
   // set mode to standby
   return(setMode(SX127X_STANDBY));
 }
 
-uint8_t SX127x::setSyncWord(uint8_t syncWord) {
+int16_t SX127x::setSyncWord(uint8_t syncWord) {
   // set mode to standby
   setMode(SX127X_STANDBY);
   
@@ -232,7 +232,7 @@ uint8_t SX127x::setSyncWord(uint8_t syncWord) {
   return(_mod->SPIsetRegValue(SX127X_REG_SYNC_WORD, syncWord));
 }
 
-uint8_t SX127x::setFrequencyRaw(float newFreq) {
+int16_t SX127x::setFrequencyRaw(float newFreq) {
   // set mode to standby
   setMode(SX127X_STANDBY);
   
@@ -241,16 +241,16 @@ uint8_t SX127x::setFrequencyRaw(float newFreq) {
   uint32_t FRF = (newFreq * (base << 19)) / 32.0;
   
   // write registers
-  uint8_t state = _mod->SPIsetRegValue(SX127X_REG_FRF_MSB, (FRF & 0xFF0000) >> 16);
+  int16_t state = _mod->SPIsetRegValue(SX127X_REG_FRF_MSB, (FRF & 0xFF0000) >> 16);
   state |= _mod->SPIsetRegValue(SX127X_REG_FRF_MID, (FRF & 0x00FF00) >> 8);
   state |= _mod->SPIsetRegValue(SX127X_REG_FRF_LSB, FRF & 0x0000FF);
   
   return(state);
 }
 
-uint8_t SX127x::config() {
+int16_t SX127x::config() {
   // set mode to SLEEP
-  uint8_t state = setMode(SX127X_SLEEP);
+  int16_t state = setMode(SX127X_SLEEP);
   if(state != ERR_NONE) {
     return(state);
   }
@@ -286,7 +286,7 @@ uint8_t SX127x::config() {
   return(state);
 }
 
-uint8_t SX127x::setMode(uint8_t mode) {
+int16_t SX127x::setMode(uint8_t mode) {
   _mod->SPIsetRegValue(SX127X_REG_OP_MODE, mode, 2, 0);
   return(ERR_NONE);
 }
